* commit '1558ea384a0d9658d173caf3ab02d72dcc9cbd85':
  feature:适配fineio读写，支持long的pos

+++ swift-base/swift-base-io/src/main/java/com/fr/swift/cube/io/impl/fineio/input/IntFineIoReader.java
@@ -27,6 +27,6 @@ public class IntFineIoReader extends BaseFineIoReader<IntBuf> implements IntRead
 
     @Override
     public int get(long pos) {
-        return FineIOAccessor.INSTANCE.getInt(readFile, (int) pos);
+        return FineIOAccessor.INSTANCE.getInt(readFile, pos);
     }
 }
